Add lint rule one component per file
