+++ .eslintrc.js
@@ -88,6 +88,12 @@ module.exports = {
       rules: {
         'import/no-unused-modules': 'off'
       }
+    },
+    {
+      files: ['**/stories/**', '**/__mocks__/**'],
+      rules: {
+        'react/no-multi-comp': 'off'
+      }
     }
   ],
   plugins: [
@@ -177,6 +183,7 @@ module.exports = {
     ],
     'react-hooks/exhaustive-deps': 'error',
     'react-hooks/rules-of-hooks': 'error',
+    'react/no-multi-comp': 'error',
     // This project doesn't use prop types
     'react/prop-types': 'off',
     'react/react-in-jsx-scope': 'off',
