+++ webview/src/experiments/components/table/body/Row.tsx
@@ -2,8 +2,9 @@ import cx from 'classnames'
 import React, { useCallback, useMemo } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
 import { EXPERIMENT_WORKSPACE_ID } from 'dvc/src/cli/dvc/contract'
-import { StubCell, CellWrapper } from './Cell'
+import { CellWrapper } from './Cell'
 import { RowContextMenu } from './RowContextMenu'
+import { StubCell } from './StubCell'
 import styles from '../styles.module.scss'
 import { RowProp } from '../../../util/interfaces'
 import { ContextMenu } from '../../../../shared/components/contextMenu/ContextMenu'
