+++ webview/src/plots/components/PathHighlight.tsx
@@ -1,0 +1,6 @@
+import React, { PropsWithChildren } from 'react'
+import styles from './styles.module.scss'
+
+export const PathHighlight: React.FC<PropsWithChildren> = ({ children }) => (
+  <span className={styles.pathHighlight}>{children}</span>
+)
