+++ webview/src/setup/components/experiments/PythonCodeBlock.tsx
@@ -1,0 +1,6 @@
+import React from 'react'
+import { CodeBlock } from '../shared/CodeBlock'
+
+export const PythonCodeBlock = ({ children }: { children: string }) => (
+  <CodeBlock language="python">{children}</CodeBlock>
+)
