+++ webview/src/shared/components/tooltip/ErrorTooltip.tsx
@@ -1,16 +1,6 @@
 import React, { ReactElement } from 'react'
-import styles from './styles.module.scss'
 import Tooltip from './Tooltip'
-import { Error } from '../icons'
-
-export const ErrorTooltipContent: React.FC<{ error?: string }> = ({
-  error
-}) => (
-  <div className={styles.errorTooltip}>
-    <Error className={styles.errorIcon} />
-    {error}
-  </div>
-)
+import { ErrorTooltipContent } from './ErrorTooltipContent'
 
 export const ErrorTooltip: React.FC<{
   error?: string
